backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
debugpy==1.6.2
Django==4.0.6
django-cas-ng==4.3.0
django-choices==1.7.2
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-etc==1.3.1
django-mdeditor==0.1.20
django-multiselectfield==0.1.12
django-oauth-toolkit==2.1.0
django-picklefield==3.0.1
django-rest-framework-nested==0.0.1
django-taggit==1.3.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==0.8
lxml==4.9.1
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-cas==1.6.0
pytz==2022.1
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
oss2==2.18.3